# 03_openapi_original.yaml
openapi: 3.1.0
info:
  title: Kaveri Stays API
  version: 1.0.0
  description: >
    Production-grade HTTP API for Kaveri Stays luxury boutique resort chain.
    Enforces role-based permissions (guest, staff, manager, owner), transactional guarantees,
    and PostgreSQL exclusion-backed booking boundaries.
servers:
  - url: http://127.0.0.1:8000
    description: Local Development Server
paths:
...
  /bookings:
    get:
      tags:
        - Bookings
      summary: List bookings
      description: Returns bookings scoped to the caller's role (guests see own, managers see property, owner sees all).
      operationId: list_bookings
      security:
        - bearerAuth: []
      parameters:
        - name: property_id
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [confirmed, checked_in, checked_out, cancelled, no_show]
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: guest_id
          in: query
          required: false
          schema:
            type: integer
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
            enum: [check_in, check_out, booking_id, guest_count]
            default: check_in
        - name: sort_order
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated bookings list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBookingsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Bookings
      summary: Create a new booking
      description: Atomically creates booking and records initial payment inside a single database transaction.
      operationId: create_booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'
        '400':
          description: Invalid request parameters (e.g. check_out <= check_in)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Room or guest not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Room already booked for the requested dates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Guest count exceeds maximum occupancy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{id}:
    get:
      tags:
        - Bookings
      summary: Get booking by ID
      operationId: get_booking_by_id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (or 404 for other guest privacy)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{id}/check-in:
    post:
      tags:
        - Bookings
      summary: Check-in a guest
      operationId: check_in_booking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Guest successfully checked in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'
        '400':
          description: Illegal state transition (booking is not confirmed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Guests cannot perform check-in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{id}/check-out:
    post:
      tags:
        - Bookings
      summary: Check-out a guest
      operationId: check_out_booking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Guest successfully checked out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'
        '400':
          description: Illegal state transition (booking is not checked_in)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Guests cannot perform check-out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/{id}/cancel:
    post:
      tags:
        - Bookings
      summary: Cancel a booking
      operationId: cancel_booking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Booking successfully cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'
        '400':
          description: Illegal state transition (cannot cancel checked_in/checked_out stays)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateBookingRequest:
      type: object
      required: [room_id, check_in, check_out, guest_count, payment_method]
      properties:
        guest_id:
          type: integer
          description: Optional for staff booking on behalf of guest; defaults to caller guest_id
        room_id:
          type: integer
          example: 1
        check_in:
          type: string
          format: date
          example: "2026-06-01"
        check_out:
          type: string
          format: date
          example: "2026-06-05"
        guest_count:
          type: integer
          example: 2
        payment_method:
          type: string
          enum: [card, upi, bank_transfer, cash]
          example: "card"
    BookingDetailResponse:
      type: object
      required: [booking_id, guest_id, room_id, check_in, check_out, guest_count, status, total_amount, amount_paid]
      properties:
        booking_id:
          type: integer
        guest_id:
          type: integer
        guest_name:
          type: string
        room_id:
          type: integer
        room_number:
          type: string
        property_name:
          type: string
        check_in:
          type: string
          format: date
        check_out:
          type: string
          format: date
        guest_count:
          type: integer
        status:
          type: string
          enum: [confirmed, checked_in, checked_out, cancelled, no_show]
        total_amount:
          type: number
          format: float
        amount_paid:
          type: number
          format: float
    PaginatedBookingsResponse:
      type: object
      required: [total, limit, offset, items]
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookingDetailResponse'
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "ROOM_UNAVAILABLE"
            message:
              type: string
              example: "The room is not available for the requested dates."
            details:
              type: array
              items:
                type: object